refactor and fix viewmodel test with navcontrol
