+++ app/src/androidTest/kotlin/com/example/returnpals/composetools/loginOptionsTest.kt
@@ -9,6 +9,9 @@ import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
+import com.example.returnpals.composetools.login.ConfirmEmailContent
+import com.example.returnpals.composetools.login.GuestLoginContent
+import com.example.returnpals.composetools.login.LoginContent
 import org.junit.Rule
 import org.junit.Test
