+++ app/src/androidTest/kotlin/com/example/returnpals/viewmodel/LoginViewModelTest.kt
@@ -1,18 +1,42 @@
 package com.example.returnpals.viewmodel
 
-import com.example.returnpals.services.LoginViewModel
+
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.navigation.compose.ComposeNavigator
+import androidx.navigation.testing.TestNavHostController
 import com.example.returnpals.dataRepository.LoginRepository
+import com.example.returnpals.navigation.AppNavigation
+import com.example.returnpals.services.LoginViewModel
 import junit.framework.TestCase.assertEquals
 import kotlinx.coroutines.async
 import kotlinx.coroutines.test.runTest
+import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
+
 
 // https://developer.android.com/kotlin/coroutines/test
 
 //Test the backend logic for the login functionality
 class LoginViewModelTest {
+    private lateinit var navController: TestNavHostController
+    private lateinit var vm: LoginViewModel
+    @get:Rule
 
-    private val vm = LoginViewModel()
+    val rule = createComposeRule()
+
+
+    @Before
+    fun setupAppNavHost() {
+        rule.setContent {
+            navController = TestNavHostController(LocalContext.current)
+            navController.navigatorProvider.addNavigator(ComposeNavigator())
+            AppNavigation(navController = navController)
+            vm = LoginViewModel("joe.biden@mail.com","p4ssw0rd", navController)
+        }
+    }
+
 
     @Test
     fun password() {
