+++ app/src/main/java/com/example/returnpals/composetools/dashboard/PickupDetails.kt
@@ -3,7 +3,7 @@ package com.example.returnpals.composetools.dashboard
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
 import com.amplifyframework.datastore.generated.model.PickupMethod
-import com.example.returnpals.composetools.goto
+import com.example.returnpals.navigation.goto
 import com.example.returnpals.composetools.pickup.PickupMethodScreen
 import com.example.returnpals.navigation.MenuRoutes
