+++ app/src/main/java/com/example/returnpals/composetools/dashboard/PickupProcess.kt
@@ -2,7 +2,7 @@ package com.example.returnpals.composetools.dashboard
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import com.example.returnpals.composetools.goto
+import com.example.returnpals.navigation.goto
 import com.example.returnpals.composetools.pickup.PickupDateScreen
 import com.example.returnpals.navigation.MenuRoutes
 import java.time.LocalDate
