+++ app/src/main/java/com/example/returnpals/navigation/Register.kt
@@ -3,7 +3,7 @@ package com.example.returnpals.navigation
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
-import com.example.returnpals.composetools.mainMenu.RegisterScreen
+import com.example.returnpals.composetools.login.RegisterScreen
 import com.example.returnpals.dataRepository.RegisterRepository
 import com.example.returnpals.viewmodel.RegisterUserInfo
 import com.example.returnpals.viewmodel.RegisterViewModel
