+++ app/src/main/java/com/example/returnpals/viewmodel/Backend.kt
@@ -9,8 +9,8 @@ import com.amplifyframework.core.Amplify
 import com.amplifyframework.datastore.AWSDataStorePlugin
 import com.amplifyframework.datastore.generated.model.Returns
 import com.amplifyframework.storage.s3.AWSS3StoragePlugin
-import com.example.returnpals.composetools.OrderRepository
-import com.example.returnpals.composetools.ProfileRepository
+import com.example.returnpals.dataRepository.OrderRepository
+import com.example.returnpals.dataRepository.ProfileRepository
 import java.io.File
 import java.io.FileOutputStream
 import java.io.InputStream
