+++ app/src/main/java/com/example/returnpals/viewmodel/OrderViewModel.kt
@@ -10,7 +10,7 @@ import com.amplifyframework.core.model.temporal.Temporal
 import com.amplifyframework.datastore.generated.model.Labels
 import com.amplifyframework.datastore.generated.model.PickupStatus
 import com.example.returnpals.PickupInfo
-import com.example.returnpals.composetools.OrderRepository
+import com.example.returnpals.dataRepository.OrderRepository
 import java.io.File
 import java.time.LocalDate
